requirements.txt
-e git+ssh://git@github.com/pierclgr/AAS_project.git@e6d04be29864479e3a82be4708155187dc083a86#egg=aas_project
ale-py==0.7.5
antlr4-python3-runtime==4.9.3
appnope==0.1.3
backcall==0.2.0
Box2D==2.3.10
certifi==2022.6.15
charset-normalizer==2.1.0
click==8.1.3
cloudpickle==1.6.0
cycler==0.11.0
decorator==4.4.2
docker-pycreds==0.4.0
ffmpeg==1.4
fonttools==4.34.4
future==0.18.2
gitdb==4.0.9
GitPython==3.1.27
glfw==2.5.4
gym==0.25.1
gym-notices==0.0.7
hydra-core==1.2.0
idna==3.3
imageio==2.21.1
imageio-ffmpeg==0.4.7
importlib-metadata==4.12.0
importlib-resources==5.9.0
iniconfig==1.1.1
ipython==7.34.0
jedi==0.18.1
kiwisolver==1.4.4
matplotlib==3.5.2
matplotlib-inline==0.1.3
moviepy==1.0.3
natsort==8.1.0
numpy==1.21.6
omegaconf==2.2.2
opencv-python==4.6.0.66
packaging==21.3
parso==0.8.3
pathtools==0.1.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.2.0
proglog==0.1.10
promise==2.3
prompt-toolkit==3.0.30
protobuf==3.20.1
psutil==5.9.1
ptyprocess==0.7.0
pygame==2.1.2
pyglet==1.5.26
Pygments==2.12.0
PyOpenGL==3.1.6
pyparsing==3.0.9
python-dateutil==2.8.2
PyYAML==6.0
requests==2.28.1
scipy==1.7.3
sentry-sdk==1.9.0
setproctitle==1.3.0
shortuuid==1.0.9
six==1.16.0
smmap==5.0.0
torch==1.12.0
tqdm==4.64.0
traitlets==5.3.0
typing_extensions==4.3.0
urllib3==1.26.11
wandb==0.12.21
wcwidth==0.2.5
zipp==3.8.1
